inline or not
