+++ src/pages/auth/SignIn.tsx
@@ -7,11 +7,10 @@ import { useSignInMutation } from '../../store/services/authApi';
 import { useAppSelector } from '../../hooks/rtk-hooks';
 import { RootState } from '../../store';
 import { useGoogleRecaptcha } from '../../hooks/useGoogleRecaptcha';
-import ItemButton from '../../widgets/buttons/ItemButton';
+import Button from '../../widgets/Button';
 import { GoSync } from 'react-icons/go';
 import { MdLogin } from 'react-icons/md';
-import ItemBadge from '../../widgets/badges/ItemBadge';
-import ContainedButton from '../../widgets/buttons/ContainedButton';
+import Badge from '../../widgets/Badge';
 
 export interface SignInForm {
   username: string;
@@ -121,21 +120,22 @@ const SignIn = () => {
         })}
       />
       {errors.password && <span>{errors.password.message}</span>}
-      <ContainedButton
+      <Button
         type="submit"
         disabled={isSubmitting}
         loading={isSubmitting}
         className="text-sm font-semibold"
+        inline
         primary
       >
         {isSubmitting ? <GoSync className="animate-spin" /> : <MdLogin />}로그인
-      </ContainedButton>
-      <ItemBadge primary pill className="text-xs">
+      </Button>
+      <Badge rounded className="text-xs">
         badge
-      </ItemBadge>
-      <ItemButton bounce warning rounded>
+      </Badge>
+      <Button bounce warning rounded>
         <MdLogin />
-      </ItemButton>
+      </Button>
       {reCaptchaElement}
     </form>
   );
